requirements/utils.txt
#utils knihovna
random-word
quote
deep-translator
googletrans
openai
django-crontab
